requirements.txt
langchain-groq
langchain_core
langchain_community
python-dotenv
pypdf
sentence-transformers
streamlit
faiss-cpu

